test: nested 를 사용한 test 관심사 분리
